use latest masstransit version for NET462
